requirements.txt
fastapi
uvicorn

sqlalchemy
pydantic

python-jose
passlib

python-multipart

python-dotenv